+++ Sources/Ukl/Src/lib/types.h
@@ -1,4 +1,4 @@
 #include <kot/types.h>
 #pragma once
 
-typedef uintptr_t pagetable_t;
+typedef uint64_t pagetable_t;
